✨ Add vulkan framework infrastructure
There are various Vulkan crates out there, but all of them are pre-1.0,
which makes it hard to trust the quality, stability, or maintenance of
the projects. Thus the decision was made to simply wrap the C API, which
is rather straight-forward to do already.

This adds a new crate for `vulkan` to handle exactly this.