🔧 Integrate `cargo-deny`
`cargo-deny` is a small and useful tool for rejecting licenses that
are not permissive enough. It's an excellent way to ensure that a project
is truly open as intended without relying on anything invalid, even
transitively.

This integrates the tool, and adds exceptions for licenses that
are transitively used only by build tooling.