✨ Add very basic `example` project
This "example" project will be the baseline to build-out the
functionality of this library. In the mean time, this will be a basic
application that can be executed and will be checked-in with new
features. However, in the future, this will be migrated to make use of
Cargo's real "example" feature.