🩹 Fix workspace "resolver" key
The `Cargo.toml` file has been generating warnings due to the
key `workspace.resolver` being set and considered redundant. This
comes from a mistake where the file was re-arranged, and the
`workspace.resolver` key got moved to below the `[workspace]` section,
resulting in `workspace.workspace.resolver` -- which is an invalid
key.

This just renames `workspace.resolver` to `resolver`, since being
under the `[workspace]` section makes this implicit and silences
the warnings.