🔨 Split C libraries into `-sys` targets
The nomenclature that appears common in Rust is to name C-libraries with
a `-sys` suffix, and to implement the Rust bindings over a separate
crate that depends on this one. Although I _personally_ find this
strange, I would rather align with the defacto standards being done
by others.

As a result, the `vulkan`, `openal`, and `glfw` crates have all been
split into `-sys` equivalents which are depended on by the Rust
bindings. This also gives the opportunity for each crate to have
different license information that aligns with the C implementations.