➕ Use `openal-soft` instead of `openal`
The base OpenAL implementation has not been maintained in a long time,
and vastly overcomplicates the rust build by forcing an installed
dependency that may vary from machine to machine.

It turns out, the often recommended OpenAL-soft implementation is not
only easy to build via CMake (nearly a carbon-copy of GLFW), but it
also was updated as recently as this past weekend -- which means it
is still actively being maintained.

This uses a git submodule to manage the OpenAL dependency, which is now
used for generation.