🔨 Build and link static libraries
The libraries were previously being compiled into dynamlic libraries and
reused from that -- which made the `*-sys` wrappers practically empty
`rlib`s with no real symbols. This makes the libraries more practical by
embedding the C binaries within it via static-linkage instead. The only
caveat to this is that the dynamic dependencies now become unfolded
manually; but one of the benefits is that the the *rlib* is now
responsible for determining _which_ library should be used to resolve
dependencies.

For example, the `glfw` _shared_ binary would rely on a specific Vulkan
implementation -- which may resolve a different instance than one
linked into a binary and used by the client (e.g. MoltenVK vs core Vulkan).
By using static linkage, the Rlibs are now responsible for resolving
these dependencies instead -- which enables a "bring your own" model
rather than an rpath enabled shared library. Ultimately this is a net
win, at the cost of convenience.

Who knows; I may change my mind in the future about this.