➕ Add ImGui framework
Imgui is being supported through the `cimgui` C-wrapper for imgui, so
that `bindgen` can be leveraged to generate proper imgui symbols for
this project. Like the binaries before this, the imgui binary is
linked statically so that the rlib can resolve dependencies. This
allows the rlib to be the one to resolve GLFW, which will resolve from
the `glfw-sys` dependency instead of an rpath potentially resolving this
as a system-installed one which may be different.

The example is updated to call these symbols as well to ensure that
it doesn't panic/crash/fatal error. No visuals are visible yet though.