✨ Add more types to `foundation` crate
The `foundation` crate has been expanded to define some additional basic
utilities for working with old C libraries, and for writing basic Rust
to begin with.

The first set of utilities are the `Transparent` utilities, which provide
a means of indicating that a type is a transparent wrapper around a C
type -- for the purpose of transfiguration between the types.

The second set of utilities are for detecting the current source
location in Rust, much like C++'s `std::source_location`.