➕ Add dependency to `boxer`
It's often necessary in engines to provide native pop-ups back to the
user; often to indicate error conditions that will soon exit.
As such, this library needs for a mechanism to accomplish this.
One simple library that doesn't reuire much setup is the Boxer library,
which is an easy MIT library that can be used; though it's authored in
C++.

This copies the Boxer sources, modifies it to have a C header, and
adds Rust bindgen bindings for the library.