🧹 Clean up cargo workspace
The `Cargo.toml` files have be suboptimally using workspaces to this
point, in that common/shared dependencies are being re-defined in each
workspace member, rather than defining it once and sharing it across
multiple `Cargo.toml` files.

This makes use of the `workspace.dependencies` Cargo workspace attribute
so that the common dependencies are only named once, so that each
deliverable `Cargo.toml` now only need reference the base version.