🔨 Reduce `glfw-sys` generated surface-area
The `glfw-sys` generated bindings were previously generating bindings
for the Vulkan API, which is undesirable since this is also generated
in `vulkan-sys` and thus creates two conflicting generation types.

It appears that `bindgen` supports this pretty easily with a `blocklist`
that can prevent generation of specific symbols; so this has been used
along with `raw_line`s that simply `use vulkan_sys::*` so that this
library can properly depend on `vulkan-sys` bindings.