🔨 Rework `Cargo.toml` dependencies
The dependencies for the `Cargo.toml` now accurately model the `feature`
settings, so that toggling `vulkan` will enable or disable all Vulkan
features the whole way down (e.g. no dep to `vulkan-sys`/`vulkan`,
no bindings for `glfw-sys` or features in `glfw`).

Additionally, the default-generated comment for the `Cargo.toml` schema
has been moved to be a consistent `Schema: `comment at the top of the
`Cargo.toml` file.