🚦 Merge Build, Test, and Coverage jobs
The jobs for build, test, and coverage were previously separate; but
this is becoming a hassle for getting the `vulkan` library to be
detected during execution. Merging this together isn't a _bad_ decision,
and will ultimately speed up CI significantly.

Instead, this will leverage `GITHUB_STEP_SUMMARY` in a future change to
make this more legible.