🚦 Add experimental promotion process
It's not, strictly-speaking, "ideal" for me to be `--force` pushing
changes to `master` to check for CI. It'd be much better to have a
process which either manually does promotion, or automates promotion
so that `master` can always be in a building state.

This makes some changes to attempt that by triggering builds on
`develop` that, on successful builds, automatically push the changes
to `master` instead. This allows a consistent `develop` branch without
the need to break `master` all the time.