🩹 Fix `GITHUB_SUMMARY_STEP` message for promote
The promotion workflow was using double-quotes instead of single-quotes
(`"` instead of `'`). This is resulting in backticks being interpreted
as a subshell instead of what it's meant to be, which is a markdown
code-span identifier. Additionally, the table format was incorrect,
which leads to unexpected appearance.

This fixes the above by using `'`, and by adding the appropriate
separator so that the table will display correctly.