🚦 Add environment deployment information
The deployment system is currently deploying any time `develop` passes
a build, with no real manual intervention. Ideally, the goal for the
future would be for these branches to synchronize only after manual
approval.

To test this out, I've added a new `master` environment which is
referenced in the CI workflow to test promotion requirements.