✨ Add meta information to `alloy` crate
The `alloy` engine's base crate needs to be able to represent itself as
a "version" so that it can be tested against, and used internally for
Vulkan-engine information. This exposes a basic semantic-`Version` type
for this purpose, and a small script to generate the Engine's version
based on the `Cargo.toml`'s specified version.