🩹 Fix `git` warnings in Github workflows
The "Checkout" action is resulting in warnings that Git is returning
nonzero exit-codes, and this is due to the lack of the `.git/module`
registration for cached 3rd-party submodules. The easy fix here is for
the cache to also include the `.git/modules` path in addition to the
existing `3rd-party` directory -- since this will enable the remote
path to be discovered and hopefully not fail.