🚦 Extract and cache code-coverge results
The code-coverage step on CI is currently one of the longest stalls on
the system, largely due to the fact that it needs to re-run tests, which
does require some level of recompilation between steps.

Since this is something that should not change if the source input
or dependencies has not changed, this adds new caching that will cache
the last generated code-coverage result and pull it back up whenever
it is needed.