🩹 Fix install path for cargo-github-annotation
The path being cached as part of the `cargo-github-annotation`
was previously incorrect, and this was resulting in the tool being
rebuilt every time rather than being picked up from the cache. The
issue stems from the missing `--root` flag, which is necessary to give
a new location for the tool to be installed to.

This updates the path accordingly.