🚦 Add doc verification to promotion
The `docs.yaml` workflow was not previously set up to be triggered
as part of the `promote` workflow. This isn't, strictly speaking,
"bad" -- but the lack of this does prevent catching workflow issues
earlier.

This modifies the `docs.yaml` workflow to now have a mode that skips
publishing that, and reuses this workflow as part of the promotion
process.