🩹 Encode runner in the `cargo-install` cache key
The cache key used for the `cargo-install` component action was a
generic string of `cargo-install-<target>`, which did not encode the
OS that the runner is running on. This appeared to be fine for a while,
but is now failing between macOS and Linux; with my suspicions being
that this is due to it allowing cross-os caching for Linux and macOS.
The docs for the cross-OS feature only calls out Windows as the
cross-target, so it's unsure if this is actually the reason.

At any rate, the fix is easy enough, and for correctness this is
what should have been done anyway.