🔨 Build `boxer-sys` with static compilation
The `boxer-sys` library was previously building and linking as a dylib,
but this is posing a challenge for linking the binary into the final
application.

In an effort to correct the issues, this is being changed into a
static binary so that the code is linked directly into the executable.