🚦 Add workflow for bundling binaries
This adds an automatic workflow that bundles binaries and releases them
as artifacts upon completion. To start with, this only does this
process for macOS binaries; with the intention of adding more in the
near future.

There is a large chance that this process will change several times
over the course of this project, especially seeing as `cargo-bundle` is
not very well fleshed-out and fails to draw values correctly from
workspace `Cargo.toml` files.