🔨 Add support for cross-compilation
The `cargo-build` composite action currently only builds for the same
system as is being built, but ideally this should allow for more than
one target system.

This adds support for it in the `rust-toolchain` and the `cargo-build`
composite modules. By default, if unspecified, this just uses the base
architecture as normal.