🔨 Add cargo build macros
The `build` crate previously contained a few `macro`s for emitting
cargo annotations, but didn't define all of them. This adds in the
remaining missing annotation macros, and added more explicit
documentation for each one.

Build scripts that were not previously using this have been updated
to make use of it.