🚦 Merge Doctest operation into build
The doctest build step was previously defined as a separate runner job,
which not only requires a duplicate build, it also is only being built
for ubuntu. This moves the stage to be part of the normal builds so that
it runs on all target platforms and guarantees that it can compile
and runs correctly.