✨ Add `phantom_fragment` macro
There are times while writing macros that it can be useful to generate
"nothing" for a fragment. Due to the way the rust macro expansion
occurs, however, this can't work without a basic utility.

Much like the `PhantomData`, the `phantom_fragment` allows unused
fragments to be "expanded" so that additional logic can be expanded
with it, even if it's not explicitly used.