🩹 Fix broken `bundle` workflow
The `bundle` workflow was broken due to some divergences is setup from
the point in which it was tested and deployed. Testing consisted of a
hard-coded profile of `dev` and a trigger from `develop` changes just
to see that it worked. In particular this should have been fine, but
it lead to a situation where the `inputs.profile` was never actually
leveraged when it should have been.

This updates the workflow to always select the appropriate profile from
the input instead, so that when the `build` workflow completes and
triggers this, it won't fail by not having a `profile` specified.