🩹 Attempt fix of `libvulkan` issue
macOS builds are failing intermittently due to not finding the
`libvulkan` binary in some circumstances -- though it is unclear why
this is happening.

Logs are showing that the `VULKAN_SDK` _does_ exist in the root and
contains `libvulkan`; just that this binary is not directly passed to
the _executable_ being built. I have an idea for what may fix it, which
is implemented here: emit the search path to include `dependency=`
_in addition to_ `native=`. Re-reading the cargo docs seemed to suggest
that `dependency` search-paths are needed when relying on a binary
transitively -- and that, by specifying `native=`, it is disabling
looking for the dependency in here by default.

This begs the separate question of why this _ever_ works in the first
place; my guess is that some macOS runners legitimately have Vulkan
installed, whereas others don't; and thus the succeeding builds found
Vulkan in `/usr/bin` or somewhere common.