🚦 Add automation for creating GH issues on failure
It is not common of libraries in Cargo to check in their `Cargo.lock`
as this pins the version for any dependents of the system. A consequence
of this is that it's always possible that an upstream dependency may
perform a breaking change without us knowing about, and this can cause
failures to build.

As a result, this now adds a CRON schedule set up nightly for this
product, and also enables automatic Github issue creation -- the logic
of which, for the time being, is taken directly from the rust-semver
project's automation, here:
https://github.com/rust-lang/rust-semverver/blob/821e059e8263e8bc76daea1a4be697954f7c2299/.github/workflows/ci.yml#L46-L76

In the future, this will migrate to a proper `actions/github-script`
instead; but for now this should work. This will also make it easier
to determine if the `libvulkan` fix actually works, or not.