✨ Add `vulkan::Result` type
Vulkan has a whole error-handling management that heavily leverages
the `VkResult` type -- which may be either an error or a truthy value
indicating that an API has not errored.

To help abstract this in a more Rust-native way, this adds a `Result`
type alias that does automatic conversions where possible.