✨ Add `Bool32` representation of `VkBool32`
Rust, unlike C or C++, does not have implicit conversions between
integer types and `bool` types. This unfortunately makes it annoying to
cast between types, where the `VkBool32` type is represented as a 32-bit
value -- but is only representing a true or false value.

To make it easier to work with, this adds a very simple type that wraps
`VkBool32`