🩹 Attempt further fix of `libvulkan` failure
The last fix didn't seem to work; however this time I was able to
identify something new with this failure: this links correctly on
_builds_, but it does not link correctly for _tests_ -- which means
that tests aren't receiving the appropriate `-L`, whereas everything
else is for building.

Clearly this means there is something off in the test setup, either
due to the `build.rs` not passing the flag, or due to the `Cargo.toml`
where the tests may for some reason be missing a required dependency (?).
The exact cause is not clear, but this is getting a few steps closer.

As an additional attempt at a fix, the `dependency=` cargo link has
been updated to be a generic link command without any type. This, by
default, becomes `all` which should always link -- which is what is
supposed to happen.