🩹 Copy `libvulkan` binary to output dir
Nightly is failing due to the same issue that has been plaguing this
project for a while, where the `vulkan-sys` project is failing to link
during tests. This finally just copies `libvulkan` when `VULKAN_SDK`
is present, since the SDK dictates what files will be available and
where. With any luck, this will finally fix the issue.

Closes: #8