🩹 Attempt to fix `libvulkan` link error again
The nightly integration tests are intermittently failing to find
`libvulkan.dylib` still on macOS. For some reason, it never includes the
required search-path to find the binary from `$VULKAN_SDK` -- even
though it should be resolvable from here.

This attempts two possible fixes for this:

* The first, is to include the `$OUT_DIR`, for which the macOS build
  currently copies the `libvulkan.dylib` binary to. This is as an attempt
  to mitigate this problem
* The second, is to use `native=`, since the other `native=` paths
  remain as `-L` in the link stages on macOS. It's not clear if this is
  just coincidental though.

It's not clear if this will do anything at all. This may just be a
transitive linking issue, though hopefully not.

Fixes: #9