✨ Add `ansi` crate for working with escape codes
Although not strictly necessary, it's often desirable in software to
support some level of color-output on the command-line, as this can
make it easier to visually wade through verbose outputs.
There is no easy way to manage this through the standard library, and
there are many non-1.0 packages in the Rust `crate.io` ecosystem. Since
this project is aiming to only depend on stable packages where possible,
I opted to write my own crate for this -- since it's easy enough.

This adds in some basic utilities for making ANSI escape codes, and offers
a simple way to integrate it with systems that identify as terminals.