✨ Add `log` package for basic logging support
Rust does not have any builtin package for basic logging, but this is
something frequently needed in videogame development -- as it can help
provide contextual details for things that occur during development.

This adds in a custom `log` package with log-sink support, complete with
ANSI formatted color output. It also updates the `example` to make use
of this package rather than still using the default builtin `println!`
messages.