🩹 Fix broken static-analysis CI job
The "Static Analysis" job appears to have been broken for a little bit,
but not reporting any failures. This is due to the use of `tee`
intercepting the compile-failure and suppressing it to a `0` status code.

This updates the workflow to setup the runner correctly so that it won't
fail; though there is still the matter of updating this job in the
future so that it won't suppress the build failures.