✨ Add `out_dir` macro to `build` crate
It's frequently necessary in `build.rs` scripts to refer to the output
location defined in the environment variable `OUT_DIR`. This is often
written verbosely as `std::env::var("OUT_DIR").unwrap()` -- which is
necessary because the ENV variable is defined _at runtime_ to the
build script (meaning `env!(...)` cannot be used).

As a means to try to clean up this readability, this is adding in a
helper function in the `build` crate simply named `build::out_dir()`
which will return the string output path that is equivalent to the
`std::env::var(...)` call.