🔨 Rename `foundation` -> `astd`
The `foundation` module was always a little bit too verbose of a crate
name to have to repeat everywhere, especially with the ubiquity of the
`cstr!` macro.

In an effort to make this easier to work with, the `foundation` crate
has been converted into a new `astd` crate and has been reorganized
to model the `std` crate structure.