🔨 Rename workflows to follow nomenclature
The various workflows were all over the place in terms of naming.
This tries to start a new nomenclature for `scheduled-` and `deploy-`
prefixes for scheduled CI runs and deployments respectively, and to use
`community-` for any community-driven workflows such as bots.

Basic CI will now be just `ci`, and other workflows lack a distinct
name (for now).

In the process, some redundant or unused trigger settings have been
dropped.