🔨 Add github release-note automation config
Github supports automatic release-note generation via a `releases.yaml`
config in the `.github` directory. It is quite possible that this will
not prove sufficient for this project, since most things do not use
Github labels -- and thus labels will be, at most, tracked via issues
and not pull-requests. THat said, this will still trial it and see how
it works.