🚦 Update to using `bitwizeshift/actions-github`
I've recently spread out into a new `actions-github` project that wraps
large amounts of the `github-script` API to simplify making composite
actions. This encourages better reuse across repositories, and also
allows for a single-source-of-truth for testing the behavior of this.

This updates to use the new repo actions, and deletes previous ones.