📌 Update `action-rust-cache` version
The `rust-cache` action is complaining for being written in Node16
instead of Node20, due to new Github deprecation notices.

It looks like the upstream `Swatinem/rust-cache` chose to do this as a
patch bump rather than the major bump that it should be. Thankfully,
since this is operating off of a fork, this allows versions to be
modeled correctly in our repository, which is being done instead.