📌 Add `mdbook-admonish` plugin support
It's frequently desirable to have sections dedicated to aside-material
like warnings, errors, tips, etc. One such case this is already a need
in the book is to prefix certain sections as not being fully
representative for the system in its current state.

To support such a case, the dependency `mdbook-admonish` has been
adopted, which leverages material design themes to accomplish this.
Some sections have been added making use of this to verify that it does
work as intended as well.