✨ Add `glew` and OpenGL support
This expands the set of supported graphical frameworks to include
OpenGL in addition to Vulkan. This is done by using `glew` to load
OpenGL functions, and wrapping a more ergonomic wrapper over top of
the raw OpenGL functions.

This initial commit only includes basic "clear" functionality, but
will expand to provide more functionality in time.