✨ Move `glew::gl` to new `gl` crate
This commit moves the `glew::gl` module to a new `gl` crate. This is
done to decouple `glew` from the `gl` instance, allowing for possible
different fulfillment of the `gl` API (such as `gl3`).

The `gl` crate is a thin wrapper around the `gl-sys` crate, which
itself is a wrapper around the fulfilling crate (`glew-sys` in this
case). This also abstracts the `__glew*` wrapper pointers into actual
calls.