🩹 Fix OpenAL CMake compilation arguments
Somehow the OpenAL CMake compilation arguments were largely specified
incorrectly, resulting in Windows options that were nonsensical and
otherwise invalid.

This corrects them. The issue occurred due to multi-line editing
of the options, and not realizing that some received a `D` prefix
due to being command-line defined as `-D` (which was used as the
baseline input when creating this), and others lost the `A` prefix from
`ALSOFT` -- likely an inverse problem from trying to cut N characters
of a prefix when some had N+1 and others had N-1.