Fix broken cargo-about composite action
When changing this over to use a dockerized action execution instead of
just a manual install, the `shell: bash` was retained by accident. A
consequence of this is that Github does not parse this as a proper
action step that uses another action -- and instead treats it like a
shell step that is otherwise broken.

This corrects it by dropping the `shell` command.