Temporarily allow diagnostic for transmute
It seems a new diagnostic for clippy was added that warns about using
transmute without explicitly stating the destination type. I don't have
the time to address this now, and I already have `SAFETY` docs on each
usage -- so I'm confident that this is not presently a problem.

This commit temporarily allows the diagnostic to be ignored until I can
address it properly.