Add OBJ model parsing module
This introduces the first version of an OBJ-format model parsing
module. This is a horribly unoptimized, but simplistic version of a
parser for `.obj` files that converts it into a raw model of data.

This will be used in future pipeline structures so that models can be
defined in a more machine-friendly format.