Add `EulerAngle` type definition
With the newly added `Quaternion` type, it can be helpful to also
have a representation of yaw, pitch, and roll angles for more
conventional representation. The current APIs provided only return
triples of angles; so this finally adds a "real" `EulerAngle` type
that identifies more clearly which angle is which.