Add lldb pretty-printers for math types
The various matrix and column types were not being displayed well in the
debugger, due to them being referential types. This adds a set of pretty
printers for the math types, which will display them in a more readable
format.

A `.lldbinit` file has been included that can easily be sourced for
users, or can be added to `launch.json` in VSCode.