Add `Transform` type
In graphics engines, it's important to have a mechanism to represent
transformations of 3d-objects -- which typically only need to model
translation, rotation, and scaling. This commit adds a `Transform` type
to the `model` module, which is a simple struct that holds a `Vec3` for
translation, a `Quaternion` for rotation, and a `Vec3` for scaling.