Add modules for constant ieee 754 operations
Floating point values in rust are not available at compile-time in
stable rust, however IEEE 754 is largely stable on all target systems.

In order to support some compile-time stable operations, this adds
an `ieee754` internal package that imports types in `core` so that we
can get some level of compile-time ops, for better optimizations.