Add hint for optimizing slice members
The generated code for `Mat4`/`Matrix4` and the various `Vec*` types
were all producing branches when checking for indices within their
guaranteed ranges. Though there is no real hint to the compiler that
does this, leveragine `std::hint::unreachable_unchecked` can be used to
tell the compiler that the branch is not needed -- and this will reduce
the code size substantially.

This adds the hint and documents the matrix and vector types with the
new hint. Some additional optimizations are rolled into the `Mat4`
implementation here to improve the overall ops and reduce unnecessary
boundary checks.