Add script for running `miri`
The rust `miri` tool is really helpful to finding undefined behavior,
since it runs all code in a small interpreter that can catch wrong
type casts, out-of-bounds accesses, and other undefined behavior.
This actually caught an issue in the matrix definition that was fixed
in one of the previous commits!

This requires the nightly tool `miri` from `rustup`, but this project
is not a nightly project. In order to make it possible to run
`miri` on CI, this adds a script that just simply calls it.