Rename "Arguments" to "Parameters" in docs
This is a minor change, but one for consistency; some areas use
"Arguments" and others use "Parameters". This change renames all
to align with "Parameters", despite this being the less-common case,
because it is more accurate of a name. Arguments typically refer
to the values passed to a function, whereas parameters refer to
the variables that store those values.