Add argument pass-through to run scripts
This adds a small but useful change to the `run-*` scripts so that
arguments passed to the script will be forwarded to the `cargo run`
command. This has been done so that it is easier to test release-builds
by passing `--release` to the script; and now also enables running
other forms of feature-builds as well.