Rerun `imgui-sys` builds on CMakeLists.txt changes
The `imgui-sys` crate was not specifying certain paths to rerun the build
script, which means it wouldn't rebuild when the source material changes.
One prominent file is the `CmakeLists.txt` file -- but this also occurs
with the OpenGL sources and the GLFW sources.

This updates the `build.rs` to account for these things.