Update CC version to `clang-18`
This project is building with `ubuntu-latest`, which updates
periodically. When this project was started, it pointed at `clang-15`
which was the only version that it included; now this is no longer
around, so we need to update the version to `clang-18` which is the
latest version available on Ubuntu runners.