Check-in Cargo.lock file
Builds have been a little inconsistent, I suspect due to some dependency
deviations between the CI and local environments. I'm going to check in
the Cargo.lock file to ensure that the CI builds are consistent with the
local builds.

This isn't often the recommended practice for libraries, but it provides
a consistent point to work from for the time being -- and this can be
re-evaluated in the future.