Disable macos builds
The macOS builds have been failing for some time for an unexplained
reason, despite building completely fine locally. For the time being,
the macOS builds will be disabled since this is the platform it's
being built on -- so it's not a priority to fix on CI.

This will be addressed out-of-band in the future.