Add `.cargo/config` to include more warnings
This enables support for detecting unused imports which can help to
clean up the codebase. Unnecessary dependencies have been removed, and
some other dependencies are kept as synthetic dependencies since they
are a "logical" requirement of the crate, but not a fundamental one.

For example, GLFW requires OpenGL -- but the OpenGL binding is actually
written using C++ instead of using Rust, so there is no formal requirement
on the OpenGl crate. However, using one without the other does not make
a lot of sense.

This helps to preserve some of this requirement.