Use `Result` for fallible `Vec*::from_*` calls
The various `Vec*` and `Vector*` constructor functions for `from_*_slice`
were previously returning `None` if the input slice was not the correct
length. This is not idiomatic Rust, and the functions should instead
return a `Result` with a reasonable error message in such cases instead.
`Result`s are also convertible into `Option` easily.