Migrate `vec*.rs` tests to separate test files
The vector files are getting way too big to be sustainable, and having
the tests merged into the same file are not helping. I don't care if this
is "idiomatic" rust; as far as I'm concerned, the idioms are wrong if
it means having to deal with overwhelmingly large test files.

From now on, this project will move towards having independen `*.test.rs`
files that are pulled in with `#[path = ...]` attributes instead. This
will cut down on the bloat and make them more visible in the file
hierarchy.