Add initial implementation of `DualQuaternion`
Dual-Quaternions help compress translations and rotations into a single
representation. This is useful for skinning and other animation
techniques where you want to interpolate between different poses.

This provides an initial implementation with some basic tests for
accuracy. However, it is also unclear whether this is correctly done --
so this might need to be revisited.