Add `Mul` and `Div` support for `Angle` trait
The `Angle` trait previously did not support `Mul` or `Div` in its
definition, which ultimately requires all uses to convert to a discrete
angle type first (e.g. `to_angle::<Radian>()`) before it could be used.

This was not ideal because it forced a conversion that may not be necessary
in the underlying angle implementation. This change now makes it so that
`Mul` and `Div` are part of the definition so that this is allowed.