Address warnings in OpenGL example
Rust has a hard time understanding output-variables from the C-API,
and are warning about unused values. If the values are not set, then it
complains about it not being initialized.

To avoid the warnings, manual `#[allow(...)]` attributes are added to
the code. This will eventually be wrapped, so this is temporary.